pyproject.toml
[project]
name = "my-first-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.6",
    "langchain-anthropic>=1.3.1",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6"
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "langgraph-cli[inmem]>=0.4.11",
    "notebook>=7.5.2",
    "pip>=25.3",
    "python-dotenv>=1.2.1",
    "streamlit>=1.55.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
